Rearrange tests to reflect the code being tested
