+++ cluster-controller/src/test/java/io/strimzi/controller/cluster/operations/resource/ResourceOperationTest.java
@@ -46,14 +46,6 @@ import io.fabric8.openshift.api.model.ImageStream;
 import io.fabric8.openshift.api.model.ImageStreamBuilder;
 import io.fabric8.openshift.client.OpenShiftClient;
 import io.fabric8.openshift.client.dsl.BuildConfigResource;
-import io.strimzi.controller.cluster.operations.resource.BuildConfigResources;
-import io.strimzi.controller.cluster.operations.resource.ConfigMapResources;
-import io.strimzi.controller.cluster.operations.resource.DeploymentResources;
-import io.strimzi.controller.cluster.operations.resource.ImageStreamResources;
-import io.strimzi.controller.cluster.operations.resource.PvcResources;
-import io.strimzi.controller.cluster.operations.resource.ResourceOperation;
-import io.strimzi.controller.cluster.operations.resource.ServiceResources;
-import io.strimzi.controller.cluster.operations.resource.StatefulSetResources;
 import io.vertx.core.Vertx;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
@@ -78,7 +70,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(VertxUnitRunner.class)
-public class ResourceOperationTest {
+public abstract class ResourceOperationTest {
 
     public static final String NAME = "name";
     public static final String NAMESPACE = "namespace";
@@ -94,13 +86,13 @@ public class ResourceOperationTest {
         vertx.close();
     }
 
-    private <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R2 extends Resource<T, D>>
+    protected <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R extends Resource<T, D>>
             void createWhenExistsIsANop(TestContext context,
-                    Class<C> clientType,
-                    Class<? extends Resource> resourceType,
-                    T resource,
-                    BiConsumer<C, MixedOperation> mocker,
-                    BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R2>> f) {
+                                        Class<C> clientType,
+                                        Class<? extends Resource> resourceType,
+                                        T resource,
+                                        BiConsumer<C, MixedOperation> mocker,
+                                        BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R>> f) {
         Resource mockResource = mock(resourceType);
         when(mockResource.get()).thenReturn(resource);
 
@@ -113,7 +105,7 @@ public class ResourceOperationTest {
         C mockClient = mock(clientType);
         mocker.accept(mockClient, mockCms);
 
-        ResourceOperation<C, T, L, D, R2> op = f.apply(vertx, mockClient);
+        ResourceOperation<C, T, L, D, R> op = f.apply(vertx, mockClient);
 
         Async async = context.async();
         op.create(resource, ar -> {
@@ -128,13 +120,13 @@ public class ResourceOperationTest {
     }
 
 
-    private <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R2 extends Resource<T, D>>
+    protected <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R extends Resource<T, D>>
             void existenceCheckThrows(TestContext context,
-                      Class<C> clientType,
-                      Class<? extends Resource> resourceType,
-                      T resource,
-                      BiConsumer<C, MixedOperation> mocker,
-                      BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R2>> f) {
+                                      Class<C> clientType,
+                                      Class<? extends Resource> resourceType,
+                                      T resource,
+                                      BiConsumer<C, MixedOperation> mocker,
+                                      BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R>> f) {
         RuntimeException ex = new RuntimeException();
 
         Resource mockResource = mock(resourceType);
@@ -149,7 +141,7 @@ public class ResourceOperationTest {
         C mockClient = mock(clientType);
         mocker.accept(mockClient, mockCms);
 
-        ResourceOperation<C, T, L, D, R2> op = f.apply(vertx, mockClient);
+        ResourceOperation<C, T, L, D, R> op = f.apply(vertx, mockClient);
 
         Async async = context.async();
         op.create(resource, ar -> {
@@ -159,13 +151,13 @@ public class ResourceOperationTest {
         });
     }
 
-    private <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R2 extends Resource<T, D>>
+    protected <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R extends Resource<T, D>>
             void successfulCreation(TestContext context,
-                    Class<C> clientType,
-                    Class<? extends Resource> resourceType,
-                    T resource,
-                    BiConsumer<C, MixedOperation> mocker,
-                    BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R2>> f) {
+                                    Class<C> clientType,
+                                    Class<? extends Resource> resourceType,
+                                    T resource,
+                                    BiConsumer<C, MixedOperation> mocker,
+                                    BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R>> f) {
         Resource mockResource = mock(resourceType);
         when(mockResource.get()).thenReturn(null);
 
@@ -178,7 +170,7 @@ public class ResourceOperationTest {
         C mockClient = mock(clientType);
         mocker.accept(mockClient, mockCms);
 
-        ResourceOperation<C, T, L, D, R2> op = f.apply(vertx, mockClient);
+        ResourceOperation<C, T, L, D, R> op = f.apply(vertx, mockClient);
 
         Async async = context.async();
         op.create(resource, ar -> {
@@ -189,13 +181,13 @@ public class ResourceOperationTest {
         });
     }
 
-    private <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R2 extends Resource<T, D>>
+    protected <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R extends Resource<T, D>>
             void creationThrows(TestContext context,
-                    Class<C> clientType,
-                    Class<? extends Resource> resourceType,
-                    T resource,
-                    BiConsumer<C, MixedOperation> mocker,
-                    BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R2>> f) {
+                                Class<C> clientType,
+                                Class<? extends Resource> resourceType,
+                                T resource,
+                                BiConsumer<C, MixedOperation> mocker,
+                                BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R>> f) {
         RuntimeException ex = new RuntimeException();
 
         Resource mockResource = mock(resourceType);
@@ -211,7 +203,7 @@ public class ResourceOperationTest {
         C mockClient = mock(clientType);
         mocker.accept(mockClient, mockCms);
 
-        ResourceOperation<C, T, L, D, R2> op = f.apply(vertx, mockClient);
+        ResourceOperation<C, T, L, D, R> op = f.apply(vertx, mockClient);
 
         Async async = context.async();
         op.create(resource, ar -> {
@@ -221,13 +213,13 @@ public class ResourceOperationTest {
         });
     }
 
-    private <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R2 extends Resource<T, D>>
+    protected <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R extends Resource<T, D>>
             void deleteWhenResourceDoesNotExistIsANop(TestContext context,
-                   Class<C> clientType,
-                   Class<? extends Resource> resourceType,
-                   T resource,
-                   BiConsumer<C, MixedOperation> mocker,
-                   BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R2>> f) {
+                                                      Class<C> clientType,
+                                                      Class<? extends Resource> resourceType,
+                                                      T resource,
+                                                      BiConsumer<C, MixedOperation> mocker,
+                                                      BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R>> f) {
         Resource mockResource = mock(resourceType);
         when(mockResource.get()).thenReturn(null);
 
@@ -240,7 +232,7 @@ public class ResourceOperationTest {
         C mockClient = mock(clientType);
         mocker.accept(mockClient, mockCms);
 
-        ResourceOperation<C, T, L, D, R2> op = f.apply(vertx, mockClient);
+        ResourceOperation<C, T, L, D, R> op = f.apply(vertx, mockClient);
 
         Async async = context.async();
         op.delete(resource.getMetadata().getNamespace(), resource.getMetadata().getName(), ar -> {
@@ -252,13 +244,13 @@ public class ResourceOperationTest {
         });
     }
 
-    private <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R2 extends Resource<T, D>>
+    protected <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R extends Resource<T, D>>
     void deleteWhenResourceExistsThrows(TestContext context,
                                         Class<C> clientType,
                                         Class<? extends Resource> resourceType,
                                         T resource,
                                         BiConsumer<C, MixedOperation> mocker,
-                                        BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R2>> f) {
+                                        BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R>> f) {
         RuntimeException ex = new RuntimeException();
 
         Resource mockResource = mock(resourceType);
@@ -273,7 +265,7 @@ public class ResourceOperationTest {
         C mockClient = mock(clientType);
         mocker.accept(mockClient, mockCms);
 
-        ResourceOperation<C, T, L, D, R2> op = f.apply(vertx, mockClient);
+        ResourceOperation<C, T, L, D, R> op = f.apply(vertx, mockClient);
 
         Async async = context.async();
         op.delete(resource.getMetadata().getNamespace(), resource.getMetadata().getName(), ar -> {
@@ -283,13 +275,13 @@ public class ResourceOperationTest {
         });
     }
 
-    private <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R2 extends Resource<T, D>>
+    protected <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R extends Resource<T, D>>
     void successfulDeletion(TestContext context,
                             Class<C> clientType,
                             Class<? extends Resource> resourceType,
                             T resource,
                             BiConsumer<C, MixedOperation> mocker,
-                            BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R2>> f) {
+                            BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R>> f) {
         Resource mockResource = mock(resourceType);
         when(mockResource.get()).thenReturn(resource);
 
@@ -302,7 +294,7 @@ public class ResourceOperationTest {
         C mockClient = mock(clientType);
         mocker.accept(mockClient, mockCms);
 
-        ResourceOperation<C, T, L, D, R2> op = f.apply(vertx, mockClient);
+        ResourceOperation<C, T, L, D, R> op = f.apply(vertx, mockClient);
 
         Async async = context.async();
         op.delete(resource.getMetadata().getNamespace(), resource.getMetadata().getName(), ar -> {
@@ -313,13 +305,13 @@ public class ResourceOperationTest {
         });
     }
 
-    private <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R2 extends Resource<T, D>>
+    protected <C extends KubernetesClient, T extends HasMetadata, L extends KubernetesResourceList, D, R extends Resource<T, D>>
     void deletionThrows(TestContext context,
                         Class<C> clientType,
                         Class<? extends Resource> resourceType,
                         T resource,
                         BiConsumer<C, MixedOperation> mocker,
-                        BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R2>> f) {
+                        BiFunction<Vertx, C, ? extends ResourceOperation<C, T, L, D, R>> f) {
         RuntimeException ex = new RuntimeException();
 
         Resource mockResource = mock(resourceType);
@@ -335,7 +327,7 @@ public class ResourceOperationTest {
         C mockClient = mock(clientType);
         mocker.accept(mockClient, mockCms);
 
-        ResourceOperation<C, T, L, D, R2> op = f.apply(vertx, mockClient);
+        ResourceOperation<C, T, L, D, R> op = f.apply(vertx, mockClient);
 
         Async async = context.async();
         op.delete(resource.getMetadata().getNamespace(), resource.getMetadata().getName(), ar -> {
@@ -345,340 +337,6 @@ public class ResourceOperationTest {
         });
     }
 
-    @Test
-    public void testConfigMapCreation(TestContext context) {
-        ConfigMap cm = new ConfigMapBuilder()
-                .withNewMetadata()
-                .withName(NAME)
-                .withNamespace(NAMESPACE)
-                .withLabels(singletonMap("foo", "bar"))
-                .endMetadata()
-                .withData(singletonMap("FOO", "BAR"))
-                .build();
 
-        BiFunction<Vertx, KubernetesClient, ConfigMapResources> f = (vertx1, client) -> new ConfigMapResources(vertx1, client);
-
-        BiConsumer<KubernetesClient, MixedOperation> mocker = (mockClient, mockCms) -> when(mockClient.configMaps()).thenReturn(mockCms);
-
-        createWhenExistsIsANop(context, KubernetesClient.class, Resource.class, cm, mocker, f);
-        existenceCheckThrows(context, KubernetesClient.class, Resource.class, cm, mocker, f);
-        successfulCreation(context, KubernetesClient.class, Resource.class, cm, mocker, f);
-        creationThrows(context, KubernetesClient.class, Resource.class, cm, mocker, f);
-    }
-
-    @Test
-    public void testDeploymentCreation(TestContext context) {
-        Deployment dep = new DeploymentBuilder().withNewMetadata().withNamespace(NAMESPACE).withName(NAME).endMetadata().build();
-        BiConsumer<KubernetesClient, MixedOperation> mocker = (mockClient, mockCms) -> {
-            ExtensionsAPIGroupDSL mockExt = mock(ExtensionsAPIGroupDSL.class);
-            when(mockExt.deployments()).thenReturn(mockCms);
-            when(mockClient.extensions()).thenReturn(mockExt);
-        };
-        BiFunction<Vertx, KubernetesClient, DeploymentResources> f = (vertx1, client) -> new DeploymentResources(vertx1, client);
-
-        createWhenExistsIsANop(context, KubernetesClient.class, ScalableResource.class, dep, mocker, f);
-        existenceCheckThrows(context, KubernetesClient.class, ScalableResource.class, dep, mocker, f);
-        successfulCreation(context, KubernetesClient.class, ScalableResource.class, dep, mocker, f);
-        creationThrows(context, KubernetesClient.class, ScalableResource.class, dep, mocker, f);
-    }
-
-    @Test
-    public void testServiceCreation(TestContext context) {
-        Service dep = new ServiceBuilder().withNewMetadata().withNamespace(NAMESPACE).withName(NAME).endMetadata().build();
-        BiConsumer<KubernetesClient, MixedOperation> mocker = (mockClient, mockCms) -> when(mockClient.services()).thenReturn(mockCms);
-
-        BiFunction<Vertx, KubernetesClient, ServiceResources> f = (vertx1, client) -> new ServiceResources(vertx1, client);
-
-        createWhenExistsIsANop(context, KubernetesClient.class, Resource.class, dep, mocker, f);
-        existenceCheckThrows(context, KubernetesClient.class, Resource.class, dep, mocker, f);
-        successfulCreation(context, KubernetesClient.class, Resource.class, dep, mocker, f);
-        creationThrows(context, KubernetesClient.class, Resource.class, dep, mocker, f);
-    }
-
-    @Test
-    public void testStatefulSetCreation(TestContext context) {
-        StatefulSet dep = new StatefulSetBuilder().withNewMetadata().withNamespace(NAMESPACE).withName(NAME).endMetadata().build();
-        BiConsumer<KubernetesClient, MixedOperation> mocker = (mockClient, mockCms) -> {
-            AppsAPIGroupDSL mockExt = mock(AppsAPIGroupDSL.class);
-            when(mockExt.statefulSets()).thenReturn(mockCms);
-            when(mockClient.apps()).thenReturn(mockExt);
-        };
-        BiFunction<Vertx, KubernetesClient, StatefulSetResources> f = (vertx1, client) -> new StatefulSetResources(vertx1, client);
-
-        createWhenExistsIsANop(context, KubernetesClient.class, RollableScalableResource.class, dep, mocker, f);
-        existenceCheckThrows(context, KubernetesClient.class, RollableScalableResource.class, dep, mocker, f);
-        successfulCreation(context, KubernetesClient.class, RollableScalableResource.class, dep, mocker, f);
-        creationThrows(context, KubernetesClient.class, RollableScalableResource.class, dep, mocker, f);
-    }
-
-    @Test
-    public void testBuildConfigCreation(TestContext context) {
-        BuildConfig dep = new BuildConfigBuilder().withNewMetadata().withNamespace(NAMESPACE).withName(NAME).endMetadata().build();
-        BiConsumer<OpenShiftClient, MixedOperation> mocker = (mockClient, mockCms) -> when(mockClient.buildConfigs()).thenReturn(mockCms);
-        BiFunction<Vertx, OpenShiftClient, BuildConfigResources> f = (vertx1, client) -> new BuildConfigResources(vertx1, client);
-
-        createWhenExistsIsANop(context, OpenShiftClient.class, BuildConfigResource.class, dep, mocker, f);
-        existenceCheckThrows(context, OpenShiftClient.class, BuildConfigResource.class, dep, mocker, f);
-        successfulCreation(context, OpenShiftClient.class, BuildConfigResource.class, dep, mocker, f);
-        creationThrows(context, OpenShiftClient.class, BuildConfigResource.class, dep, mocker, f);
-    }
-
-    @Test
-    public void testImageStreamCreation(TestContext context) {
-        ImageStream dep = new ImageStreamBuilder().withNewMetadata().withNamespace(NAMESPACE).withName(NAME).endMetadata().build();
-        BiConsumer<OpenShiftClient, MixedOperation> mocker = (mockClient, mockCms) -> when(mockClient.imageStreams()).thenReturn(mockCms);
-        BiFunction<Vertx, OpenShiftClient, ImageStreamResources> f = (vertx1, client) -> new ImageStreamResources(vertx1, client);
-
-        createWhenExistsIsANop(context, OpenShiftClient.class, Resource.class, dep, mocker, f);
-        existenceCheckThrows(context, OpenShiftClient.class, Resource.class, dep, mocker, f);
-        successfulCreation(context, OpenShiftClient.class, Resource.class, dep, mocker, f);
-        creationThrows(context, OpenShiftClient.class, Resource.class, dep, mocker, f);
-    }
-
-    @Test
-    public void testConfigMapDeletion(TestContext context) {
-        ConfigMap cm = new ConfigMapBuilder()
-                .withNewMetadata()
-                .withName(NAME)
-                .withNamespace(NAMESPACE)
-                .withLabels(singletonMap("foo", "bar"))
-                .endMetadata()
-                .withData(singletonMap("FOO", "BAR"))
-                .build();
-
-        BiFunction<Vertx, KubernetesClient, ConfigMapResources> f = (vertx1, client) -> new ConfigMapResources(vertx1, client);
-
-        BiConsumer<KubernetesClient, MixedOperation> mocker = (mockClient, mockCms) -> when(mockClient.configMaps()).thenReturn(mockCms);
-        deleteWhenResourceDoesNotExistIsANop(context,
-                KubernetesClient.class,
-                Resource.class,
-                cm,
-                mocker,
-                f);
-        deleteWhenResourceExistsThrows(context,
-                KubernetesClient.class,
-                Resource.class,
-                cm,
-                mocker,
-                f);
-        successfulDeletion(context,
-                KubernetesClient.class,
-                Resource.class,
-                cm,
-                mocker,
-                f);
-        deletionThrows(context,
-                KubernetesClient.class,
-                Resource.class,
-                cm,
-                mocker,
-                f);
-    }
-
-    @Test
-    public void testDeploymentDeletion(TestContext context) {
-        Deployment dep = new DeploymentBuilder().withNewMetadata().withNamespace(NAMESPACE).withName(NAME).endMetadata().build();
-        BiConsumer<KubernetesClient, MixedOperation> mocker = (mockClient, mockCms) -> {
-            ExtensionsAPIGroupDSL mockExt = mock(ExtensionsAPIGroupDSL.class);
-            when(mockExt.deployments()).thenReturn(mockCms);
-            when(mockClient.extensions()).thenReturn(mockExt);
-        };
-
-        BiFunction<Vertx, KubernetesClient, DeploymentResources> f = (vertx1, client) -> new DeploymentResources(vertx1, client);
-
-        deleteWhenResourceDoesNotExistIsANop(context,
-                KubernetesClient.class,
-                ScalableResource.class,
-                dep,
-                mocker,
-                f);
-        deleteWhenResourceExistsThrows(context,
-                KubernetesClient.class,
-                ScalableResource.class,
-                dep,
-                mocker,
-                f);
-        successfulDeletion(context,
-                KubernetesClient.class,
-                ScalableResource.class,
-                dep,
-                mocker,
-                f);
-        deletionThrows(context,
-                KubernetesClient.class,
-                ScalableResource.class,
-                dep,
-                mocker,
-                f);
-    }
-
-    @Test
-    public void testServiceDeletion(TestContext context) {
-        Service dep = new ServiceBuilder().withNewMetadata().withNamespace(NAMESPACE).withName(NAME).endMetadata().build();
-        BiFunction<Vertx, KubernetesClient, ServiceResources> f = (vertx1, client) -> new ServiceResources(vertx1, client);
-        BiConsumer<KubernetesClient, MixedOperation> mocker = (mockClient, mockCms) -> {
-            when(mockClient.services()).thenReturn(mockCms);
-        };
-
-        deleteWhenResourceDoesNotExistIsANop(context,
-                KubernetesClient.class,
-                ScalableResource.class,
-                dep,
-                mocker,
-                f);
-        deleteWhenResourceExistsThrows(context,
-                KubernetesClient.class,
-                ScalableResource.class,
-                dep,
-                mocker,
-                f);
-        successfulDeletion(context,
-                KubernetesClient.class,
-                ScalableResource.class,
-                dep,
-                mocker,
-                f);
-        deletionThrows(context,
-                KubernetesClient.class,
-                ScalableResource.class,
-                dep,
-                mocker,
-                f);
-    }
-
-    @Test
-    public void testStatefulSetDeletion(TestContext context) {
-        StatefulSet dep = new StatefulSetBuilder().withNewMetadata().withNamespace(NAMESPACE).withName(NAME).endMetadata().build();
-        BiFunction<Vertx, KubernetesClient, StatefulSetResources> f = (vertx1, client) -> new StatefulSetResources(vertx1, client);
-        BiConsumer<KubernetesClient, MixedOperation> mocker = (mockClient, mockCms) -> {
-            AppsAPIGroupDSL mockExt = mock(AppsAPIGroupDSL.class);
-            when(mockExt.statefulSets()).thenReturn(mockCms);
-            when(mockClient.apps()).thenReturn(mockExt);
-        };
-
-        deleteWhenResourceDoesNotExistIsANop(context,
-                KubernetesClient.class,
-                RollableScalableResource.class,
-                dep,
-                mocker,
-                f);
-        deleteWhenResourceExistsThrows(context,
-                KubernetesClient.class,
-                RollableScalableResource.class,
-                dep,
-                mocker,
-                f);
-        successfulDeletion(context,
-                KubernetesClient.class,
-                RollableScalableResource.class,
-                dep,
-                mocker,
-                f);
-        deletionThrows(context,
-                KubernetesClient.class,
-                RollableScalableResource.class,
-                dep,
-                mocker,
-                f);
-    }
-
-    @Test
-    public void testPvcDeletion(TestContext context) {
-        PersistentVolumeClaim dep = new PersistentVolumeClaimBuilder().withNewMetadata().withNamespace(NAMESPACE).withName(NAME).endMetadata().build();
-        BiFunction<Vertx, KubernetesClient, PvcResources> f = (vertx1, client) -> new PvcResources(vertx1, client);
-        BiConsumer<KubernetesClient, MixedOperation> mocker = (mockClient, mockCms) -> {
-            when(mockClient.persistentVolumeClaims()).thenReturn(mockCms);
-        };
-
-        deleteWhenResourceDoesNotExistIsANop(context,
-                KubernetesClient.class,
-                Resource.class,
-                dep,
-                mocker,
-                f);
-        deleteWhenResourceExistsThrows(context,
-                KubernetesClient.class,
-                Resource.class,
-                dep,
-                mocker,
-                f);
-        successfulDeletion(context,
-                KubernetesClient.class,
-                Resource.class,
-                dep,
-                mocker,
-                f);
-        deletionThrows(context,
-                KubernetesClient.class,
-                RollableScalableResource.class,
-                dep,
-                mocker,
-                f);
-    }
-
-    @Test
-    public void testBuildConfigDeletion(TestContext context) {
-        BuildConfig dep = new BuildConfigBuilder().withNewMetadata().withNamespace(NAMESPACE).withName(NAME).endMetadata().build();
-        BiFunction<Vertx, OpenShiftClient, ResourceOperation<OpenShiftClient, BuildConfig, BuildConfigList, DoneableBuildConfig, BuildConfigResource<BuildConfig, DoneableBuildConfig, Void, Build>>> f = (vertx1, client) -> new BuildConfigResources(vertx1, client);
-        BiConsumer<OpenShiftClient, MixedOperation> mocker = (mockClient, mockCms) ->
-                when(mockClient.buildConfigs()).thenReturn(mockCms);
-
-        deleteWhenResourceDoesNotExistIsANop(context,
-                OpenShiftClient.class,
-                BuildConfigResource.class,
-                dep,
-                mocker,
-                f);
-        deleteWhenResourceExistsThrows(context,
-                OpenShiftClient.class,
-                BuildConfigResource.class,
-                dep,
-                mocker,
-                f);
-        successfulDeletion(context,
-                OpenShiftClient.class,
-                BuildConfigResource.class,
-                dep,
-                mocker,
-                f);
-        deletionThrows(context,
-                OpenShiftClient.class,
-                BuildConfigResource.class,
-                dep,
-                mocker,
-                f);
-    }
-
-    @Test
-    public void testImageStreamDeletion(TestContext context) {
-        ImageStream dep = new ImageStreamBuilder().withNewMetadata().withNamespace(NAMESPACE).withName(NAME).endMetadata().build();
-        BiFunction<Vertx, OpenShiftClient, ImageStreamResources> f = (vertx1, client) -> new ImageStreamResources(vertx1, client);
-        BiConsumer<OpenShiftClient, MixedOperation> mocker = (mockClient, mockCms) ->
-                when(mockClient.imageStreams()).thenReturn(mockCms);
-
-        deleteWhenResourceDoesNotExistIsANop(context,
-                OpenShiftClient.class,
-                Resource.class,
-                dep,
-                mocker,
-                f);
-        deleteWhenResourceExistsThrows(context,
-                OpenShiftClient.class,
-                Resource.class,
-                dep,
-                mocker,
-                f);
-        successfulDeletion(context,
-                OpenShiftClient.class,
-                Resource.class,
-                dep,
-                mocker,
-                f);
-        deletionThrows(context,
-                OpenShiftClient.class,
-                Resource.class,
-                dep,
-                mocker,
-                f);
-    }
 }
+
